Move secp256k1 test
