+++ programs/secp256k1/Cargo.toml
@@ -13,6 +13,12 @@ publish = false
 
 [dependencies]
 solana-sdk = { path = "../../sdk", version = "1.4.0" }
+libsecp256k1 = "0.3.5"
+sha3 = "0.9.1"
+digest = "0.9.0"
+bincode = "1.3.1"
+rand = "0.7.0"
+solana-logger = { path = "../../logger", version = "1.4.0" }
 
 [lib]
 crate-type = ["lib"]
